+++ src/core/mod.rs
@@ -1,5 +1,6 @@
 pub mod integrity;
 pub mod nu_version;
+pub mod official_registry;
 pub mod package;
 pub mod platform;
 pub mod registry;